Fix protoc generating warning.
Generated file name "Private\SGreeter\GreeterService_Private.h" does not conform to the Protobuf generation specification. Note that the file name must be relative, use "/" instead of "", and not use "." or ".." as part of the file name. Buf will continue without error here, but please raise an issue with the maintainer of the plugin and reference